+++ src/triangulation/mod.rs
@@ -1,7 +1,9 @@
 mod cdt;
 mod delaunay2;
 mod ids;
+mod polygon;
 
 pub use cdt::{CdtError, ConstrainedTriangulation2, constrained_delaunay2, validate_cdt_topology};
 pub use delaunay2::{Triangulation2, delaunay2};
 pub use ids::{EdgeId, FaceId, VertexId};
+pub use polygon::{PolygonTriangulationError, triangulate_polygon};